chore: Seperate anime and manga functions, ready to start working on manga
