+++ components/anime/layout/Popular.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { getAnimePopular } from "../../../src/handlers";
+import { getAnimePopular } from "../../../src/handlers/anime";
 import AnimeCard from "../../anime/AnimeCard";
 
 const Popular = () => {
