+++ pages/anime/info/[animeId].js
@@ -4,7 +4,7 @@ import AnimeDetails from "../../../components/anime/info/AnimeDetails";
 import MainLayout from "../../../components/ui/MainLayout";
 import Recommendations from "../../../components/anime/info/Recommendations";
 import TextButton from "../../../components/buttons/TextButton";
-import { getAnimeDetails, getAnimeEpisodeData } from "../../../src/handlers/index";
+import { getAnimeDetails, getAnimeEpisodeData } from "../../../src/handlers/anime";
 
 export const getServerSideProps = async (context) => {
   const { animeId } = context.query;
